+++ iceoryx_posh/source/popo/ports/server_port_data.cpp
@@ -21,11 +21,10 @@ namespace iox
 {
 namespace popo
 {
-/// @todo iox-#27 remove when the port queue policies are consolidated
-cxx::VariantQueueTypes getRequestQueueType(const QueueFullPolicy2 policy) noexcept
+cxx::VariantQueueTypes getRequestQueueType(const QueueFullPolicy policy) noexcept
 {
-    return policy == QueueFullPolicy2::DISCARD_OLDEST_DATA ? cxx::VariantQueueTypes::SoFi_MultiProducerSingleConsumer
-                                                           : cxx::VariantQueueTypes::FiFo_MultiProducerSingleConsumer;
+    return policy == QueueFullPolicy::DISCARD_OLDEST_DATA ? cxx::VariantQueueTypes::SoFi_MultiProducerSingleConsumer
+                                                          : cxx::VariantQueueTypes::FiFo_MultiProducerSingleConsumer;
 }
 
 ServerPortData::ServerPortData(const capro::ServiceDescription& serviceDescription,
@@ -34,12 +33,9 @@ ServerPortData::ServerPortData(const capro::ServiceDescription& serviceDescripti
                                mepoo::MemoryManager* const memoryManager,
                                const mepoo::MemoryInfo& memoryInfo) noexcept
     : BasePortData(serviceDescription, runtimeName, serverOptions.nodeName)
-    , m_chunkSenderData(memoryManager,
-                        static_cast<SubscriberTooSlowPolicy>(serverOptions.clientTooSlowPolicy),
-                        HISTORY_REQUEST_OF_ZERO,
-                        memoryInfo)
+    , m_chunkSenderData(memoryManager, serverOptions.clientTooSlowPolicy, HISTORY_REQUEST_OF_ZERO, memoryInfo)
     , m_chunkReceiverData(getRequestQueueType(serverOptions.requestQueueFullPolicy),
-                          static_cast<QueueFullPolicy>(serverOptions.requestQueueFullPolicy))
+                          serverOptions.requestQueueFullPolicy)
     , m_offeringRequested(serverOptions.offerOnCreate)
 {
     m_chunkReceiverData.m_queue.setCapacity(serverOptions.requestQueueCapacity);