+++ iceoryx_posh/test/moduletests/test_popo_chunk_sender.cpp
@@ -107,9 +107,9 @@ class ChunkSender_test : public Test
     ChunkQueueData_t m_chunkQueueData{iox::popo::QueueFullPolicy::DISCARD_OLDEST_DATA,
                                       iox::cxx::VariantQueueTypes::SoFi_SingleProducerSingleConsumer};
     ChunkSenderData_t m_chunkSenderData{
-        &m_memoryManager, iox::popo::SubscriberTooSlowPolicy::DISCARD_OLDEST_DATA, 0}; // must be 0 for test
+        &m_memoryManager, iox::popo::ConsumerTooSlowPolicy::DISCARD_OLDEST_DATA, 0}; // must be 0 for test
     ChunkSenderData_t m_chunkSenderDataWithHistory{
-        &m_memoryManager, iox::popo::SubscriberTooSlowPolicy::DISCARD_OLDEST_DATA, HISTORY_CAPACITY};
+        &m_memoryManager, iox::popo::ConsumerTooSlowPolicy::DISCARD_OLDEST_DATA, HISTORY_CAPACITY};
 
     iox::popo::ChunkSender<ChunkSenderData_t> m_chunkSender{&m_chunkSenderData};
     iox::popo::ChunkSender<ChunkSenderData_t> m_chunkSenderWithHistory{&m_chunkSenderDataWithHistory};