+++ iceoryx_posh/test/moduletests/test_popo_publisher_port.cpp
@@ -81,7 +81,7 @@ class PublisherPort_test : public Test
 
     // publisher port w/o offer on create
     iox::popo::PublisherOptions m_noOfferOnCreatePublisherOptions{
-        0U, iox::NodeName_t{""}, false, iox::popo::SubscriberTooSlowPolicy::DISCARD_OLDEST_DATA};
+        0U, iox::NodeName_t{""}, false, iox::popo::ConsumerTooSlowPolicy::DISCARD_OLDEST_DATA};
     iox::popo::PublisherPortData m_publisherPortDataNoOfferOnCreate{
         iox::capro::ServiceDescription("a", "b", "c"), "myApp", &m_memoryManager, m_noOfferOnCreatePublisherOptions};
     iox::popo::PublisherPortRouDi m_sutNoOfferOnCreateRouDiSide{&m_publisherPortDataNoOfferOnCreate};
@@ -89,7 +89,7 @@ class PublisherPort_test : public Test
 
     // publisher port that waits for subscriber when queue is full
     iox::popo::PublisherOptions m_waitForSubscriberPublisherOptions{
-        0U, iox::NodeName_t{""}, false, iox::popo::SubscriberTooSlowPolicy::WAIT_FOR_SUBSCRIBER};
+        0U, iox::NodeName_t{""}, false, iox::popo::ConsumerTooSlowPolicy::WAIT_FOR_CONSUMER};
     iox::popo::PublisherPortData m_publisherPortDataWaitForSubscriber{
         iox::capro::ServiceDescription("a", "b", "c"), "myApp", &m_memoryManager, m_waitForSubscriberPublisherOptions};
     iox::popo::PublisherPortRouDi m_sutWaitForSubscriberRouDiSide{&m_publisherPortDataWaitForSubscriber};
@@ -97,7 +97,7 @@ class PublisherPort_test : public Test
 
     // publisher port w/ history
     iox::popo::PublisherOptions m_withHistoryPublisherOptions{
-        iox::MAX_PUBLISHER_HISTORY, iox::NodeName_t{""}, true, iox::popo::SubscriberTooSlowPolicy::DISCARD_OLDEST_DATA};
+        iox::MAX_PUBLISHER_HISTORY, iox::NodeName_t{""}, true, iox::popo::ConsumerTooSlowPolicy::DISCARD_OLDEST_DATA};
     iox::popo::PublisherPortData m_publisherPortDataHistory{
         iox::capro::ServiceDescription("x", "y", "z"), "myApp", &m_memoryManager, m_withHistoryPublisherOptions};
     iox::popo::PublisherPortUser m_sutWithHistoryUserSide{&m_publisherPortDataHistory};
@@ -133,7 +133,7 @@ TEST_F(PublisherPort_test, noWaitingForSubscriberWithDefaultOptions)
 {
     ::testing::Test::RecordProperty("TEST_ID", "d1f74874-257a-4e8f-aabf-8eadad5b4367");
     EXPECT_THAT(m_sutWithDefaultOptionsRouDiSide.getOptions().subscriberTooSlowPolicy,
-                Eq(iox::popo::SubscriberTooSlowPolicy::DISCARD_OLDEST_DATA));
+                Eq(iox::popo::ConsumerTooSlowPolicy::DISCARD_OLDEST_DATA));
 }
 
 TEST_F(PublisherPort_test, initialStateReturnsOfferCaProMessageWithDefaultOptions)
@@ -158,7 +158,7 @@ TEST_F(PublisherPort_test, waitingForSubscriberWhenDesired)
 {
     ::testing::Test::RecordProperty("TEST_ID", "49526d1a-e81a-4e4a-8fb4-1a96dee83ae7");
     EXPECT_THAT(m_sutWaitForSubscriberRouDiSide.getOptions().subscriberTooSlowPolicy,
-                Eq(iox::popo::SubscriberTooSlowPolicy::WAIT_FOR_SUBSCRIBER));
+                Eq(iox::popo::ConsumerTooSlowPolicy::WAIT_FOR_CONSUMER));
 }
 
 TEST_F(PublisherPort_test, offerCallResultsInOfferedState)